[Modify] PartyId Type int to long modification

+++ src/main/java/com/example/moonkey/domain/Party.java
@@ -24,7 +24,7 @@ public class Party {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="partyId")
-    private int partyId;
+    private long partyId;
 
     @OneToMany
     @JoinColumn(name = "accountUid")
